requirements.txt
dataclasses
ensure
fastapi[standard]
google-genai
inngest
ipykernel
ipywidgets
langchain-core
langchain-openai
langchain
python-box
python-dotenv
qdrant-client
fastapi-cors
guardrails-ai == 0.6.7
langgraph
langchain-mcp-adapters
langchain-community